Move stuff from globe_control_utils to globe_utils
