+++ src/ui/handler_inertia.ts
@@ -4,7 +4,7 @@ import {bezier, clamp, extend} from '../util/util';
 import Point from '@mapbox/point-geometry';
 import type {DragPanOptions} from './handler/shim/drag_pan';
 import {EaseToOptions} from './camera';
-import {computeGlobePanCenter} from './globe_control_utils';
+import {computeGlobePanCenter} from '../geo/projection/globe_utils';
 
 const defaultInertiaOptions = {
     linearity: 0.3,
